Plugin fixed
At long last, I fixed the plugin and it now works as it should. Please report any sentence logic errors that occur so I can fix them.

The problem, in the end, was that I put a piece of code somewhere it shouldn't have been and killed the message.

Also better chat 5.2.10 is now required as opposed to 5.2.5 due to Rust updates.